requirements.txt
Django~=3.1
django-cors-headers~=3.6.0
djangorestframework~=3.11.2
djangorestframework-simplejwt~=4.6.0
django-filter~=2.4.0
mysqlclient~=2.0.1
drf-yasg~=1.20.0